in case of broken EC2 instance submitted - silently fail the update
This can be caused if the EC2 instance for which the notification submitted is already dead and has no IP address - so we have no idea which record to remove. The previous attempt to send an empty change batch didn't go well with Route53.

Ideally, we have the ASG name so we can detect this error and go look at the ASG then sanitize the DNS record by removing all entries that are no longer valid. Maybe for the next version.